+++ packages/typespec-rust/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * Adds `PagerState` support to pagers.
 * Adds a `Format` to `RequestContent<T, F>` when the request body is not JSON.
+* Calls `Context::to_borrowed` instead of `Context::with_context`.
 
 ## 0.19.0 (2025-07-23)